Completed Excercise: Revealing Content  and...
Fixed form to match the form according to course video.
NOTE: In Google Chrome browser (ver. 55.0.2883.87 m) at 1920x1080
resolution, the Cancel button in the modal appears to be in the next
line, but at 90% zoom, it is inline with Submit button. No such
behaviour noted in Microsoft Edge or Microsoft Internet Explorer.